requirements.txt
awscli==1.18.142
boto3==1.15.1
botocore==1.18.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.1
dnspython==2.0.0
docutils==0.15.2
ecdsa==0.14.1
Flask==1.1.2
Flask-PyMongo==2.3.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
pyasn1==0.4.8
pycparser==2.20
pymongo==3.11.0
python-dateutil==2.8.1
python-jose==3.2.0
PyYAML==5.3.1
requests==2.24.0
rsa==4.5
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
waitress==1.4.4
Werkzeug==1.0.1
